pyproject.toml
[tool.poetry]
name = "vision-compare"
version = "1.0.0"
description = "A project to compare performant computer vision neural network models"

license = "MIT"
authors = ["Andras Geiszl"]

readme = "README.md"
repository = "https://github.com/geiszla/vision-compare"
homepage = "https://github.com/geiszla/vision-compare"

[tool.poetry.dependencies]
python = "~3.7"
easydict = "^1.9"
keras = "^2.3.1"
matplotlib = "^3.2.1"
nptyping = "^1.0.1"
numpy = "^1.18.3"
opencv-python = "4.1.0.25"
Pillow = "^7.1.2"
requests = "^2.23.0"
tabulate = "^0.8.7"
tensorflow-gpu = "^1.15.2"
tqdm = "^4.45.0"

keras_retinanet = { path = "./lib/keras_retinanet" }

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pylint = "^2.4.4"
